+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 ## 0.66.0 (Unreleased)
 
 FEATURES:
+* mdb: Kafka topic management via dedicated resource `yandex_mdb_kafka_topic` is implemented
 * mdb: support update `name`, `description`, `labels`, `config.access.data_lens`, `config.access.web_sql`, `deletion_protection` in `yandex_mdb_greenplum_cluster` resource
 * mdb: add `access` and `backup_window_start` entities in `yandex_mdb_greenplum_cluster` resource and data source
 * mdb: support MongoDB 5.0 in `yandex_mdb_mongodb_cluster`
 * mdb: add `backup_retain_period_days` attribute to `config` entity in `yandex_mdb_postgresql_cluster` resource and data source
 
 ENHANCEMENTS:
-* mdb: skip topic management if unmanaged_topics is ON
 * add docs on k8s node group placement groups.
 * ydb: increase the default timeout to create serverless database
 * add `name`, `replication_source`, `replication_source_name` attributes to `host` entity in `yandex_mdb_mysql_cluster` resource and data source